# lesson12/ClinicService/ClinicService/Models/Requests/UpdatePetRequest.cs
// Pet/Requests/UpdatePetRequest.cs
namespace ClinicService.Models.Requests
{
    public class UpdatePetRequest
    {
        public int PetId { get; set; }
        public int ClientId { get; set; }
        public string Name { get; set; }
        public DateTime Birthday { get; set; }
    }
}